refactor(exception): separate ViewResolver from exceptionHandler
